requirements.txt
asgiref==3.7.2
Django==5.0
django-cors-headers==4.3.1
django-rest-framework==0.1.0
djangorestframework==3.14.0
imageio==2.33.0
lazy_loader==0.3
networkx==3.2.1
nibabel==5.1.0
numpy==1.26.2
packaging==23.2
Pillow==10.1.0
python-decouple==3.8
pytz==2023.3.post1
scikit-image==0.22.0
scipy==1.11.4
sqlparse==0.4.4
tifffile==2023.9.26
typing_extensions==4.8.0
tzdata==2023.3
